fix (assert): remove printf type warning

+++ include/pmsis/rtos/pmsis_assert.h
@@ -21,7 +21,7 @@
 
 #define pi_assert(test) \
 if (!(test)) { \
-printf("PMSIS assertion error in %s, %s:%ld :\n",__func__, __FILE__, __LINE__); \
+printf("PMSIS assertion error in %s, %s:%d :\n",__func__, __FILE__, (unsigned int)__LINE__); \
 printf("%s\n", #test); \
 pmsis_exit(-1); \
 }
